Update README.md
Electronic Medical Record is a web-based application which is used to capture, organize, maintain and retrieve user’s medical records on day-to-day basis. This application consists of a comprehensive database used to store and access patient's healthcare information. The major goal of EMR is to serve as a repository for prescriptions, test reports and daily health related events.